only expose an address name when a child terminated to prevent trying to send messages to it
